fix: move the environment/render-profile contract into a dependency-free module so client CI can import it (#4630)
The client parity test imported server/lib/threejsModel.js, which imports zod —
a package the client CI job does not install, so the suite failed to resolve the
module at all (it passes locally only because the workspaces share node_modules).
The constants and the resolver have no dependencies, so they move to
server/lib/threejsModelEnvironment.js and threejsModel.js consumes them, matching
how client/src/utils/providers.test.js already reaches aiToolkit/constants.js.
